+++ src/main/java/com/supermartijn642/rechiseled/data/RechiseledTextureProvider.java
@@ -16,15 +16,15 @@ public class RechiseledTextureProvider extends ChiseledTextureProvider {
 
     @Override
     protected void createTextures(){
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/acacia_planks"), "block/acacia_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/birch_planks"), "block/birch_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/crimson_planks"), "block/crimson_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/dark_oak_planks"), "block/dark_oak_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/jungle_planks"), "block/jungle_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/mangrove_planks"), "block/mangrove_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/spruce_planks"), "block/spruce_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/warped_planks"), "block/warped_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/bamboo_planks"), "block/bamboo_planks");
-        this.createPlankTextures(new ResourceLocation("minecraft", "block/cherry_planks"), "block/cherry_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/acacia_planks"), "block/acacia_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/birch_planks"), "block/birch_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/crimson_planks"), "block/crimson_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/dark_oak_planks"), "block/dark_oak_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/jungle_planks"), "block/jungle_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/mangrove_planks"), "block/mangrove_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/spruce_planks"), "block/spruce_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/warped_planks"), "block/warped_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/bamboo_planks"), "block/bamboo_planks");
+        this.createPlankTextures(ResourceLocation.fromNamespaceAndPath("minecraft", "block/cherry_planks"), "block/cherry_planks");
     }
 }